// api/src/utils/respostas/codigos-resposta.d.ts
export declare const CODIGOS_SUCESSO: {
    SUCESSO_GENERICO: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_LISTAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_EXISTE_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_N_EXISTE_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_CRIAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_EDITAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_EXCLUIR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_LISTAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_EXISTE_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_N_EXISTE_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_CRIAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_EDITAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_EXCLUIR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_LISTAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_EXISTE_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_N_EXISTE_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CRIAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_EDITAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_EDITAR_FOTOS_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_EXCLUIR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_LISTAR_FOTOS_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_LISTAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_EXISTE_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_N_EXISTE_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_CRIAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_EDITAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_EXCLUIR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_LISTAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_EXISTE_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_N_EXISTE_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_CRIAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_EDITAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_EXCLUIR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_LISTAR_SUCCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_EXISTE_SUCCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_N_EXISTE_SUCCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_CRIAR_SUCCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_EDITAR_SUCCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_EXCLUIR_SUCCESS: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_CORRESPONDE_SUCCESS: {
        codigo: number;
        mensagem: string;
    };
    TIPO_MATERIAL_LISTAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    UNIDADE_MEDIDA_LISTAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    RELATORIO_LISTAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CATEGORIA_LISTAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CATEGORIA_CRIAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CATEGORIA_EDITAR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CATEGORIA_EXCLUIR_SUCESS: {
        codigo: number;
        mensagem: string;
    };
};